Updating README and package dependency instructions
This repo was created due to OpenAI API changes. I think in the future it can be better to create a frozen version of the requirements so that it will always work in the future.

I got it by creating a virtual env and freezing the environment.

Also the requirements.txt was missing a dependency.